+++ packages/typera-express/src/context.ts
@@ -1,10 +1,10 @@
 import express = require('express')
 
-export interface ExpressContext {
+export interface RequestBase {
   req: express.Request
   res: express.Response
 }
 
-export function getRouteParams(e: ExpressContext): any {
+export function getRouteParams(e: RequestBase): any {
   return e.req.params
 }