+++ packages/typera-koa/src/middleware.ts
@@ -1,14 +1,14 @@
 import * as commonMiddleware from 'typera-common/middleware'
 import * as commonResponse from 'typera-common/response'
-import { KoaRequestBase } from './context'
+import { RequestBase } from './context'
 
 export type Middleware<
   Result,
   Response extends commonResponse.Generic
-> = commonMiddleware.Middleware<KoaRequestBase, Result, Response>
+> = commonMiddleware.Middleware<RequestBase, Result, Response>
 
 export type Generic = commonMiddleware.Middleware<
-  KoaRequestBase,
+  RequestBase,
   any,
   commonResponse.Generic
 >